Move all file requests/responses into file.rs
